1028hxf

+++ backEnd/src/main/java/nnu/mnr/satellite/model/dto/resources/SceneImageDTO.java
@@ -38,7 +38,7 @@ public class SceneImageDTO {
     @TableField(value = "bands", typeHandler = SetTypeHandler.class)
     private HashSet<String> bands;
 
-    private Integer cloud;
+    private float cloud;
     private String cloudPath;
     private String bucket;
     private String resolution;
