+++ backEnd/src/main/java/nnu/mnr/satellite/model/po/resources/Tile.java
@@ -27,7 +27,7 @@ public class Tile {
     private String tileId;
     private String tileLevel;
     private String imageId;
-    private Integer cloud;
+    private float cloud;
     private String band;
     private Integer columnId;
     private Integer rowId;
